fixed compile error. *blush*

+++ mainlib/videoreader.cpp
@@ -132,7 +132,7 @@ void VideoReader::createImageForFrameNumber(RealLifeVideo& rlv, const qreal dist
 {
     // the first few frames are sometimes black, so when requested to take a "screenshot" of the first frames, just
     // skip to a few frames after the start.
-    qint64 frameNumber = qMax(5, rlv.frameForDistance(distance));
+    qint64 frameNumber = qMax(5, static_cast<int>(rlv.frameForDistance(distance)));
 
     qDebug() << "creating image for" << rlv.name() << "distance" << distance << "frame nr" << frameNumber;
     av_seek_frame(_formatContext, _currentVideoStream, frameNumber, AVSEEK_FLAG_FRAME | AVSEEK_FLAG_BACKWARD);
